Moved class to ui package
